solve login bugs
