--- src/Login/actiontypes.js
@@ -1,5 +1,0 @@
-export const LOGIN = "LOGIN";
-export const FAILURE_LOGIN = "FAILURE_LOGIN";
-export const SUCCESS_LOGIN = "SUCCESS_LOGIN";
-export const START_LOGIN = "START_LOGIN";
-export const END_LOGIN = "END_LOGIN";
