+++ src/actions/actionTypes.js
@@ -1,3 +1,8 @@
 export const SIGNUP_LOAD = "SIGNUP_LOAD";
 export const SIGNUP_SUCCESS = "SIGNUP_SUCCESS";
 export const SIGNUP_FAILURE = "SIGNUP_FAILURE";
+export const LOGIN = "LOGIN";
+export const FAILURE_LOGIN = "FAILURE_LOGIN";
+export const SUCCESS_LOGIN = "SUCCESS_LOGIN";
+export const START_LOGIN = "START_LOGIN";
+export const END_LOGIN = "END_LOGIN";
