+++ src/reducers/index.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import { signupReducer } from './signupReducer'; 
+import { loginReducer } from './loginReducer.js';
 
 export default combineReducers({
-    signupReducer
+    signupReducer, loginReducer
   });
