+++ src/views/MentorMe.js
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
-import Signup from "../components/Signup";
+import Signup from "../components/Signup.js";
+import Login from "../components/Login.js";
 
 
 export default class MentorMe extends Component {
@@ -7,6 +8,7 @@ export default class MentorMe extends Component {
         return (
             <div>
                 <Signup />
+                <Login />
             </div>
         )
     }
